flask_requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
tensorflow==2.15.0
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3


